requirements.txt
python-dotenv~=1.0.1
SQLAlchemy~=2.0.32
requests~=2.32.3
aiohttp~=3.10.3